fix la norme fr this time

+++ inc/builtins.h
@@ -6,7 +6,7 @@
 /*   By: maroy <maroy@student.42.qc>                        ██ ██             */
 /*                                                          ██ ███████.qc     */
 /*   Created: 2023/09/01 14:30:21 by maroy                                    */
-/*   Updated: 2023/09/26 13:49:34 by maroy            >(.)__ <(.)__ =(.)__    */
+/*   Updated: 2023/10/05 15:55:07 by maroy            >(.)__ <(.)__ =(.)__    */
 /*                                                     (___/  (___/  (___/    */
 /* ************************************************************************** */
 
@@ -91,7 +91,7 @@ uint8_t		exit_builtin(char **args);
  * utils.c *
  */
 
-int16_t		env_count(void);
+int			env_count(void);
 void		modify_env(char *arg, char *key);
 void		sort_and_print(void);
 void		free_old_env(char **env_pointer);
